Poll for keyboard and mouse updates
Before the poller threads would just check for updates every 500ms. Now we indefinitely block until an update occurs and then wait 500ms after notifying the condvar to limit the rate of updates.

Note - This whole loop should be streamlined so we have one main thread polling on the device event fds and block to fade in/out depending on whether there are events or not. We don't need to have three threads always waiting around.